No borrarte a ti mismo

+++ webApp/controllers/gestion_usuarios_controller.js
@@ -269,7 +269,8 @@ exports.postModUser = (request, response) => {
                                     roles : roles,
                                     terapeuta : terapeuta,
                                     tBool : tBool,
-                                    permisos: request.session.permisos
+                                    permisos: request.session.permisos,
+                                    user: request.session.user
                                 })
                             })
                             .catch((err) => {
@@ -284,7 +285,8 @@ exports.postModUser = (request, response) => {
                             usuarios : usuarios,
                             roles : roles,
                             tBool : tBool,
-                            permisos: request.session.permisos
+                            permisos: request.session.permisos,
+                            user: request.session.user
                         });
                     }
                 })
@@ -543,49 +545,18 @@ exports.postUpdateUser = (request, response) => {
 exports.postDeleteUser = (request, response) => {
     let oldEmail = request.body.oldEmail2;
     let tBool = request.body.tBool2;
+    let user = request.body.user;
 
-    if (tBool == 'true') {
-        Grupos_Terapeutas.fetchIfTerapeutaHaveGroups(oldEmail)
-            .then(([numGrupos]) => {
-                if (numGrupos[0].num_groups == 0){
-                    Terapeuta.deleteById(oldEmail)
-                        .then(() => {
-                            Usuario_Rol.deleteById(oldEmail)
-                                .then(() => {
-                                    Usuario.deleteById(oldEmail)
-                                        .then(() => {
-                                            request.session.mensaje = 'El usuario fue eliminado';
-                                            request.session.bandera = false; 
-                                            response.redirect('/gestionAdmin/gestionUsuarios');
-                                        })
-                                        .catch((err) => {
-                                            request.session.mensaje = 'Error de comunicacion con el servidor';
-                                            request.session.bandera = true;
-                                            response.redirect('/gestionAdmin/gestionUsuarios');
-                                            console.log(err);
-                                        })
-                                })
-                                .catch((err) => {
-                                    request.session.mensaje = 'Error de comunicacion con el servidor';
-                                    request.session.bandera = true;
-                                    response.redirect('/gestionAdmin/gestionUsuarios');
-                                    console.log(err);
-                                })
-                        })
-                        .catch((err) => {
-                            request.session.mensaje = 'Error de comunicacion con el servidor';
-                            request.session.bandera = true;
-                            response.redirect('/gestionAdmin/gestionUsuarios');
-                            console.log(err);
-                        })
-                }
-                else {
-                    Usuario_Rol.deleteById(oldEmail)
-                        .then(() => {
-                            Terapeuta.changeStatusToB(oldEmail)
-                                .then(() => {
-                                    let timestamp = new Date().getUTCMilliseconds();
-                                        Usuario.changeLogin(oldEmail, timestamp)
+    if (user != oldEmail) {
+        if (tBool == 'true') {
+            Grupos_Terapeutas.fetchIfTerapeutaHaveGroups(oldEmail)
+                .then(([numGrupos]) => {
+                    if (numGrupos[0].num_groups == 0){
+                        Terapeuta.deleteById(oldEmail)
+                            .then(() => {
+                                Usuario_Rol.deleteById(oldEmail)
+                                    .then(() => {
+                                        Usuario.deleteById(oldEmail)
                                             .then(() => {
                                                 request.session.mensaje = 'El usuario fue eliminado';
                                                 request.session.bandera = false; 
@@ -597,13 +568,70 @@ exports.postDeleteUser = (request, response) => {
                                                 response.redirect('/gestionAdmin/gestionUsuarios');
                                                 console.log(err);
                                             })
-                                })
-                                .catch((err) => {
-                                    request.session.mensaje = 'Error de comunicacion con el servidor';
-                                    request.session.bandera = true;
-                                    response.redirect('/gestionAdmin/gestionUsuarios');
-                                    console.log(err);
-                                })
+                                    })
+                                    .catch((err) => {
+                                        request.session.mensaje = 'Error de comunicacion con el servidor';
+                                        request.session.bandera = true;
+                                        response.redirect('/gestionAdmin/gestionUsuarios');
+                                        console.log(err);
+                                    })
+                            })
+                            .catch((err) => {
+                                request.session.mensaje = 'Error de comunicacion con el servidor';
+                                request.session.bandera = true;
+                                response.redirect('/gestionAdmin/gestionUsuarios');
+                                console.log(err);
+                            })
+                    }
+                    else {
+                        Usuario_Rol.deleteById(oldEmail)
+                            .then(() => {
+                                Terapeuta.changeStatusToB(oldEmail)
+                                    .then(() => {
+                                        let timestamp = new Date().getUTCMilliseconds();
+                                            Usuario.changeLogin(oldEmail, timestamp)
+                                                .then(() => {
+                                                    request.session.mensaje = 'El usuario fue eliminado';
+                                                    request.session.bandera = false; 
+                                                    response.redirect('/gestionAdmin/gestionUsuarios');
+                                                })
+                                                .catch((err) => {
+                                                    request.session.mensaje = 'Error de comunicacion con el servidor';
+                                                    request.session.bandera = true;
+                                                    response.redirect('/gestionAdmin/gestionUsuarios');
+                                                    console.log(err);
+                                                })
+                                    })
+                                    .catch((err) => {
+                                        request.session.mensaje = 'Error de comunicacion con el servidor';
+                                        request.session.bandera = true;
+                                        response.redirect('/gestionAdmin/gestionUsuarios');
+                                        console.log(err);
+                                    })
+                            })
+                            .catch((err) => {
+                                request.session.mensaje = 'Error de comunicacion con el servidor';
+                                request.session.bandera = true;
+                                response.redirect('/gestionAdmin/gestionUsuarios');
+                                console.log(err);
+                            })
+                    }
+                })
+                .catch((err) => {
+                    request.session.mensaje = 'Error de comunicacion con el servidor';
+                    request.session.bandera = true;
+                    response.redirect('/gestionAdmin/gestionUsuarios');
+                    console.log(err);
+                })
+        }
+        else {
+            Usuario_Rol.deleteById(oldEmail)
+                .then(() => {
+                    Usuario.deleteById(oldEmail)
+                        .then(() => {
+                            request.session.mensaje = 'El usuario fue eliminado';
+                            request.session.bandera = false; 
+                            response.redirect('/gestionAdmin/gestionUsuarios');
                         })
                         .catch((err) => {
                             request.session.mensaje = 'Error de comunicacion con el servidor';
@@ -611,37 +639,19 @@ exports.postDeleteUser = (request, response) => {
                             response.redirect('/gestionAdmin/gestionUsuarios');
                             console.log(err);
                         })
-                }
-            })
-            .catch((err) => {
-                request.session.mensaje = 'Error de comunicacion con el servidor';
-                request.session.bandera = true;
-                response.redirect('/gestionAdmin/gestionUsuarios');
-                console.log(err);
-            })
+                })
+                .catch((err) => {
+                    request.session.mensaje = 'Error de comunicacion con el servidor';
+                    request.session.bandera = true;
+                    response.redirect('/gestionAdmin/gestionUsuarios');
+                    console.log(err);
+                })
+        }
     }
     else {
-        Usuario_Rol.deleteById(oldEmail)
-            .then(() => {
-                Usuario.deleteById(oldEmail)
-                    .then(() => {
-                        request.session.mensaje = 'El usuario fue eliminado';
-                        request.session.bandera = false; 
-                        response.redirect('/gestionAdmin/gestionUsuarios');
-                    })
-                    .catch((err) => {
-                        request.session.mensaje = 'Error de comunicacion con el servidor';
-                        request.session.bandera = true;
-                        response.redirect('/gestionAdmin/gestionUsuarios');
-                        console.log(err);
-                    })
-            })
-            .catch((err) => {
-                request.session.mensaje = 'Error de comunicacion con el servidor';
-                request.session.bandera = true;
-                response.redirect('/gestionAdmin/gestionUsuarios');
-                console.log(err);
-            })
+        request.session.mensaje = 'No puedes borrarte a ti mismo';
+        request.session.bandera = true;
+        response.redirect('/gestionAdmin/gestionUsuarios');
     }
 };
